Move extern/nmrglue.py to plugin package
- Move vendored nmrglue (2059 lines, Bruker binary reader) to
  spectrochempy-topspin plugin since it's only used by read_topspin
- Update plugin imports from spectrochempy.extern.nmrglue to local .nmrglue
- Remove nmrglue from extern/__init__.pyi stub
