+++ plugins/spectrochempy-topspin/src/spectrochempy_topspin/nmrglue.py
