+++ plugins/spectrochempy-topspin/src/spectrochempy_topspin/read_topspin.py
@@ -36,9 +36,10 @@ from spectrochempy.core.dataset.coord import Coord
 from spectrochempy.core.readers.importer import Importer
 from spectrochempy.core.readers.importer import _importer_method
 from spectrochempy.core.units import ur
-from spectrochempy.extern.nmrglue import read_fid
-from spectrochempy.extern.nmrglue import read_pdata
 from spectrochempy.utils.meta import Meta
+
+from .nmrglue import read_fid
+from .nmrglue import read_pdata
 
 # ======================================================================================
 # Constants
